Update Likes functions
