+++ src/index.js
@@ -1,8 +1,8 @@
 import './styles/main.css';
 import displayBook from './modules/displayBooks.js';
-import displayLikes from './modules/displayLikes.js';
+import updateLikes from './modules/updateLikes.js';
 
 window.addEventListener('load', () => {
   displayBook();
-  displayLikes();
+  updateLikes();
 });
