+++ src/modules/urlAndId.js
@@ -1,0 +1,4 @@
+const apiUrl = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi';
+const appId = 'xsMgySSOLdWVdMRKDgjB';
+
+export { appId, apiUrl };
